+++ CHANGELOG.md
@@ -8,4 +8,5 @@
 
 ### Maintenance
 
+- 2026-09-19 | 🔧 chore(server): validate Vapor bootstrap
 - 2026-09-19 | 🔧 chore: initialize iOS and Vapor workspace